Allow filleting outer edge of the model
Creates circular corners, with the radius specified in surface units (mm).
To ensure the resulting mesh is valid, the radius cannot be larger than the
`base_height` so that even zero-height GeoTIFF points will be rounded to a
height above the bottom of the model.

Code really starting to creak under the weight of feature additions, likely
to need a better organisational principle soon.